Refactor Done

+++ Assets/Scripts/UI and Human Interaction/NoMB/GameSnapshot.cs
@@ -15,7 +15,7 @@ public sealed class GameSnapshot
     public int[] pieceOwner;
     public byte[] pieceType;
     public short[] pieceHP;
-    public byte?[] connector;
+    public ushort?[] connector;
 
     // --- Match counters ---
     public int centerVP;
